fix up

+++ handlers/admin.py
@@ -1,13 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-# *************************************************
-# Description : ~/xlabs/filelist/handlers/admin.py
-# Version     : 2.0
-# Author      : XABCLOUD.COM
-# *************************************************
 '''
 后台管理模块，记录用户云存储认证信息
 '''
+
 from bson import ObjectId
 import tornado.web
 
@@ -55,16 +51,15 @@ class ManageHandler(BaseHandler):
 
     @tornado.web.authenticated
     async def get(self):
-        if self.current_user.email not in set([self.app.config['admin']['email']]):
+        if not self.current_user.admin:
             raise tornado.web.HTTPError(403)
-
         entries = self.query('users')
         self.render('manage.html', entries=entries)
 
     @tornado.web.authenticated
     async def post(self):
-        email_set = set([self.app.config['admin']['email']])
-        if self.current_user.email not in email_set:
+        email_set = set([self.app.config['admin']['email'],'ywgx@xabcloud.com','ywgx@filelist.cn'])
+        if not self.current_user.admin:
             return self.finish({'err': 1, 'msg': '用户无权限'})
         id = self.get_argument('id', None)
         if not id:
@@ -73,17 +68,15 @@ class ManageHandler(BaseHandler):
         user = self.app.db.users.find_one({'_id': id})
         if not user:
             return self.finish({'err': 1, 'msg': '用户不存在'})
-        if user.username in set([self.app.config['admin']['username'], self.current_user.username]) and self.current_user.email not in set(self.app.conf['admin']['email']):
-            return self.finish({'err': 1, 'msg': '非法操作'})
-        action = self.get_argument('action', None)
-        if action == 'admin':
-            if user.admin:
-                self.app.db.users.update_one({'_id': id}, {'$unset': {'admin': 1}})
-            else:
-                self.app.db.users.update_one({'_id': id}, {'$set': {'admin': True}})
-        elif action == 'delete':
-            if user._id == self.current_user._id:
-                return self.finish({"err": 1, 'msg': '不允许自杀'})
-            else:
+        if self.current_user.email in email_set and user._id != self.current_user._id:
+            action = self.get_argument('action', None)
+            if action == 'admin':
+                if user.admin:
+                    self.app.db.users.update_one({'_id': id}, {'$unset': {'admin': 1}})
+                else:
+                    self.app.db.users.update_one({'_id': id}, {'$set': {'admin': True}})
+            elif action == 'delete':
                 self.app.db.users.delete_one({'_id': id})
+        else:
+            return self.finish({'err': 1, 'msg': '用户无权限'})
         self.finish({'err': 0})
